refactor delete

+++ temp.c
@@ -3,13 +3,14 @@
 #include "menu.h"
 #include "tasks.h"
 #include "lists.h"
+#include "helpers.h"
 
 // TODO: Implement error handling for file operations in the markTaskDone and deleteTask functions. Currently, if there is an error opening the files, the program simply prints an error message and continues execution. It would be better to handle these errors more gracefully, such as by exiting the function early or providing feedback to the user about the failure.
 // TODO: Consider adding functionality to allow users to mark tasks as not done or to edit existing tasks. This would enhance the usability of the task manager and provide more flexibility for users to manage their tasks effectively../
-// delete does not work
 
 int main(void)
 {   
+    // listName format: "listName.txt"
     char listName[101];
     int exit = 0;
     int task;
@@ -24,12 +25,22 @@ int main(void)
         {    
         case 1: 
                 clearScreen();
-                promptUserForListName(listName, sizeof(listName));
+                promptUserForListName(listName, sizeof listName);
+                if (checkIfListExists(listName))
+                {
+                    printf("List already exists.\n");
+                    break;
+                }
                 addList(listName);
                 break;
         case 2: 
                 clearScreen();
-                promptUserForListName(listName, sizeof(listName));
+                promptUserForListName(listName, sizeof listName);
+                if (!checkIfListExists(listName))
+                {
+                    printf("List does not exist.\n");
+                    break;
+                }
                 deleteList(listName);
                 // if (succes == 0)
                 // {
@@ -43,7 +54,7 @@ int main(void)
         case 3: 
                 int previousMenu = 0;
                 clearScreen();
-                promptUserForListName(listName, sizeof(listName));
+                promptUserForListName(listName, sizeof listName);
                 if (!checkIfListExists(listName))
                 {
                         break;
